delete file if exixt, new apk

+++ server/src/helper/deleteFile.js
@@ -4,8 +4,17 @@ const logger = require("./winstonLogger");
 const deleteFile = async (filePath, defaultFilePath = "") => {
   try {
     if (filePath !== defaultFilePath) {
-      await fs.access(filePath);
-      await fs.unlink(filePath);
+      // Check if the file exists 
+      const isFileExists = await fs
+        .access(filePath)
+        .then(() => true)
+        .catch(() => false);
+
+      if (isFileExists) {
+        // delete the file
+        await fs.unlink(filePath);
+      }
+
       logger.info("File deleted successfully");
     } else {
       logger.info("File can't be deleted");
